Important principle shift: don't expect the exact data type
Add number boxing

This is important because it is sugar. As you know sugar always
results in diabete. This commit makes things easier, but more complex,
and a bit slower.

When type isn't strictly enforced, and when use can't choose to be
more lenient about it, it allows subtle bugs which are difficult to
remove. A better approach would be to keep providing strong certainty
to the user about data types (as it is currently), but also allow some
more relaxed behaviours: for exemple a dev settings could accept
nonetheless to ease the boxing but issue a warning. An unsafe mode
would cast silently.

Number boxing takes time to process

This leads me to think the current code structure is wrong. Instead of
using multimethods to apply operations on types, each schema should be
preprocessed (be it just in time then memoised) with code structure
similar to middlewares / interceptors. So that playing with coercion
would be a matter of adding / reordering interceptors, instead of
extending various multimethods with numerous implementation.

This preprocessing step should result in a black box (fn, macro?) of
optimised code which convert records and clj datastructures back and
forth.

Also, naming should be thoroughly rethought.

Also, a good reading of the code of metosin/malli may prove
influential in the good direction.

Also, the purpose piotr-yuxuan/dove, this library's sister, should be
archived and I should open a PR against malli to extend the schema
translation to avro schema.