Explicit, reliable way to manage union encoding
It gives the user the ability to say what it wants. It no longer
applies a Tinder strategy (see link) which swallows exceptions and
just try to encode the datum in the first possible type. Swallowing
exceptions is an anti-pattern, and in the previous implementation the
use had no way to explicit state wether `{:my-field 1}` should be
wrapped into an Avro map, or a record. This implemented strategy is
opinionated, and you are welcome to challenge it.

https://github.com/piotr-yuxuan/slava/blob/0bdc79327403afe84375ddc65522684a75ba838a/src/main/clojure/com/slava/clj%3C_%3Eavro.clj#L250

Regarding code style and not the strategy: yes it is uncommon, but it
reduces nested indentations and it's been very fun to design. After
all it's 11.30pm on a Sunday evening and this is a personal project,
so feel free either not to use it, either to say thank you and not to
complain. Cheers 😘