Avro-idiomatic logical type management
This uses a mutable, global, shared object. This is three enormous
reason to loath that verbose code. However I believe it is the best
way to integrate around Avro and make sure that user is no indered by
this ancillary library.

Plus, obviously I really like to throw piles of untested, generated
code to git, create a releases, and see how things go :/ Of course
tests are coming.